Add rudimentary support for improveTouch CLI
The original hbtp daemon had a CLI that can be accessed with
a file rename to "hbtp_tool" or running with option of "tool".

This implementation uses linenoise, a BSD licensed utility, to
create the required CLI. There are a few important commands
supported, namely "sendCommand".

Features to be implemented include logging and recording, but
these could be ignored for now.

Change-Id: I692d4e47b617240b10df76c6a95b0081cf4c9095