Wire up to prepare for enabling logging features
In order for enablement of logging, setAuth() needs to be called on
initialization with some custom magic. The set-up provided here is
functional, but the code style leads much to be desired.

Nonetheless, it is now possible to get a log telling the client that
enabling recording failed as a wrong recording parameter is provided.
This needs to be added to the TouchCompanionClient class.

Change-Id: I155d1b6422f7792ebd9b541f664d35de388c00c9